Add BasicActor for simple Erlang-like Actors
