+++ Gemfile
@@ -2,6 +2,7 @@ source 'https://rubygems.org'
 gemspec
 
 gem 'coveralls', require: false
+gem 'pry'
 
 if RUBY_PLATFORM =~ /darwin/
   gem 'rb-fsevent', '~> 0.9.1'
