+++ lib/celluloid.rb
@@ -452,6 +452,7 @@ require 'celluloid/links'
 require 'celluloid/logger'
 require 'celluloid/mailbox'
 require 'celluloid/method'
+require 'celluloid/handlers'
 require 'celluloid/receivers'
 require 'celluloid/registry'
 require 'celluloid/responses'
@@ -469,6 +470,7 @@ require 'celluloid/proxies/async_proxy'
 require 'celluloid/proxies/future_proxy'
 require 'celluloid/proxies/block_proxy'
 
+require 'celluloid/basic_actor'
 require 'celluloid/actor'
 require 'celluloid/future'
 require 'celluloid/pool_manager'
