+++ lib/celluloid/handlers.rb
@@ -1,0 +1,42 @@
+require 'set'
+
+module Celluloid
+  class Handlers
+    def initialize
+      @handlers = Set.new
+    end
+
+    def handle(*patterns, &block)
+      patterns.each do |pattern|
+        handler = Handler.new pattern, block
+        @handlers << handler
+      end
+    end
+
+    # Handle incoming messages
+    def handle_message(message)
+      handler = @handlers.find { |h| h.match(message) }
+      return unless handler
+
+      handler.call message
+      true
+    end
+  end
+
+  # Methods blocking on a call to receive
+  class Handler
+    def initialize(pattern, block)
+      @pattern = pattern
+      @block = block
+    end
+
+    # Match a message with this receiver's block
+    def match(message)
+      @pattern === message
+    end
+
+    def call(message)
+      @block.call message
+    end
+  end
+end
